Refactoring graph.cpp & graph.hpp. Added graph_sizing to support constants. Created small graph test.
